requirements.txt
requests
python-dotenv
yfinance
beautifulsoup4
anthropic
fastapi
uvicorn[standard]
